requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.1

# GitHub Integration
PyGithub==2.1.1

# Authentication
authlib==1.3.0
requests==2.31.0

# Scheduling
schedule==1.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Type Hints
typing-extensions==4.9.0

# Environment Variables
python-dotenv==1.0.0
